requirements.txt
PyQt5>=5.15.0
pytubefix>=2.2.0
youtube-transcript-api>=0.6.0
python-docx>=0.8.11
yt-dlp>=2023.3.4
browser-cookie3>=0.19.1
requests>=2.28.2
beautifulsoup4>=4.11.2
matplotlib>=3.7.1
sounddevice>=0.4.6
numpy>=1.24.3
scipy>=1.10.1
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
PyPDF2>=2.10.0
reportlab>=3.6.8
pillow>=9.0.0
selenium>=4.10.0
webdriver-manager>=4.0.0
soundfile>=0.12.1
gTTS>=2.3.2
# playsound>=1.3.0 # Removed due to installation issues
pyttsx3>=2.90 # For accessing system voices 